// lib/generated/inputTypeSchemas/SpaUncheckedUpdateInputSchema.ts
import type { Prisma } from '../../prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { SpaUpdatevideosInputSchema } from './SpaUpdatevideosInputSchema';
import { SpaUpdateimagesInputSchema } from './SpaUpdateimagesInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { SpaPriceCategoryUncheckedUpdateManyWithoutSpaNestedInputSchema } from './SpaPriceCategoryUncheckedUpdateManyWithoutSpaNestedInputSchema';

export const SpaUncheckedUpdateInputSchema: z.ZodType<Prisma.SpaUncheckedUpdateInput> = z.object({
  id: z.union([ z.string().uuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  name_esp: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  description: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  description_esp: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  hotelId: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  type: z.union([ z.string(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  videos: z.union([ z.lazy(() => SpaUpdatevideosInputSchema),z.string().array() ]).optional(),
  images: z.union([ z.lazy(() => SpaUpdateimagesInputSchema),z.string().array() ]).optional(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  updatedAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  pricecategory: z.lazy(() => SpaPriceCategoryUncheckedUpdateManyWithoutSpaNestedInputSchema).optional()
}).strict();

export default SpaUncheckedUpdateInputSchema;